backend/requirements.txt
# FastAPI and Server
fastapi==0.109.0
uvicorn[standard]==0.27.0

# Database - PostgreSQL
sqlalchemy==2.0.25
asyncpg==0.29.0
psycopg2-binary==2.9.9

# Data Validation
pydantic==2.5.3
pydantic-settings==2.1.0

# GraphQL
strawberry-graphql[fastapi]==0.219.0

# Neo4j
neo4j==5.16.0

# Redis
redis==5.0.1

# Utilities
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.26.0
pytest-cov==4.1.0
